Ported datalog code to be compatible with windows compiler. No functional changes made. Re-ran unit test and it passed. [git-p4: depot-paths = "//depot/main/embedded/Common/": change = 14784]

+++ datalog/datalog_buffer.cpp
@@ -1,8 +1,10 @@
 /*
  * Copyright (C) 2002 Gambro BCT, Inc.  All rights reserved.
  *
- * $Header: //bctquad3/home/BCT_Development/vxWorks/Common/datalog/rcs/datalog_buffer.cpp 1.7 2003/11/10 17:46:05Z jl11312 Exp rm70006 $
+ * $Header: I:/BCT_Development/vxWorks/Common/datalog/rcs/datalog_buffer.cpp 1.7 2003/11/10 17:46:05Z jl11312 Exp $
  * $Log: datalog_buffer.cpp $
+ * Revision 1.7  2003/11/10 17:46:05Z  jl11312
+ * - corrections from data log unit tests (see IT 6598)
  * Revision 1.6  2003/02/25 16:10:07Z  jl11312
  * - modified buffering scheme to help prevent buffer overruns
  *
@@ -10,7 +12,6 @@
 
 #include "datalog.h"
 #include "datalog_internal.h"
-#include "error.h"
 
 bool DataLog_BufferManager::createChain(DataLog_BufferChain & chain, unsigned long reserveBuffers)
 {
@@ -161,7 +162,7 @@ void DataLog_BufferManager::modifyChainData(DataLog_BufferChain & chain, unsigne
 		currentBufferPos += DataLog_BufferSize;
 	}
 
-   int readPos = 0;
+   size_t readPos = 0;
    int writePos = offset%DataLog_BufferSize;
 
 	while ( buffer &&
